refactor: split motor control out of main.c into modules
Extract behavior-neutral control path units so demag, faults, and
port work can land without editing a 2500-line god file:

- commutation.c: getBemfState, commutate, advanceincrement, zcfoundroutine
- bemf_zc.c: PeriodElapsedCallback, interruptRoutine, startMotor
- control_loop.c: setInput, tenKhzRoutine, PID/current helpers
- motor_runtime.h: extern map for shared runtime state still defined in main

F051 duty/COM timer macros use statement form for GCC 15 -Werror.
Builds verified: ARK_4IN1_F051 (HWCI_PERF=1) and SEQURE_4IN1_F421.
