+++ Inc/bemf_zc.h
@@ -1,0 +1,11 @@
+/*
+ * bemf_zc.h - zero-cross detection and commutation scheduling
+ */
+#ifndef BEMF_ZC_H_
+#define BEMF_ZC_H_
+
+void PeriodElapsedCallback(void);
+void interruptRoutine(void);
+void startMotor(void);
+
+#endif /* BEMF_ZC_H_ */
