+++ Mcu/f051/Inc/peripherals.h
@@ -16,14 +16,21 @@
 #define RELOAD_WATCHDOG_COUNTER() (LL_IWDG_ReloadCounter(IWDG))
 #define DISABLE_COM_TIMER_INT() (COM_TIMER->DIER &= ~((0x1UL << (0U))))
 #define ENABLE_COM_TIMER_INT() (COM_TIMER->DIER |= (0x1UL << (0U)))
-#define SET_AND_ENABLE_COM_INT(time)                                  \
-    (COM_TIMER->CNT = 0, COM_TIMER->ARR = time, COM_TIMER->SR = 0x00, \
-        COM_TIMER->DIER |= (0x1UL << (0U)))
+#define SET_AND_ENABLE_COM_INT(time) do { \
+    COM_TIMER->CNT = 0; \
+    COM_TIMER->ARR = (time); \
+    COM_TIMER->SR = 0x00; \
+    COM_TIMER->DIER |= (0x1UL << (0U)); \
+} while (0)
 #define SET_INTERVAL_TIMER_COUNT(intertime) (INTERVAL_TIMER->CNT = intertime)
 #define SET_PRESCALER_PWM(presc) (TIM1->PSC = presc)
 #define SET_AUTO_RELOAD_PWM(relval) (TIM1->ARR = relval)
-#define SET_DUTY_CYCLE_ALL(newdc) \
-    (TIM1->CCR1 = newdc, TIM1->CCR2 = newdc, TIM1->CCR3 = newdc)
+/* Statement form avoids -Wunused-value on comma-operator expansions (GCC 12+). */
+#define SET_DUTY_CYCLE_ALL(newdc) do { \
+    TIM1->CCR1 = (newdc); \
+    TIM1->CCR2 = (newdc); \
+    TIM1->CCR3 = (newdc); \
+} while (0)
 
 void initAfterJump(void);
 void initCorePeripherals(void);
