+++ Src/main.c
@@ -228,6 +228,11 @@ an settings option)
 #include "serial_telemetry.h"
 #include "kiss_telemetry.h"
 #include "hwci_perf.h"
+#include "commutation.h"
+#include "bemf_zc.h"
+#include "control_loop.h"
+
+/* Control path: commutation.c, bemf_zc.c, control_loop.c (split from main). */
 #include "signal.h"
 #include "sounds.h"
 #include "targets.h"
@@ -595,31 +600,6 @@ uint8_t ubAnalogWatchdogStatus = RESET;
 volatile char input_ready = 0;
 #endif
 
-int32_t doPidCalculations(struct fastPID* pidnow, int actual, int target)
-{
-
-    pidnow->error = actual - target;
-    pidnow->integral = pidnow->integral + pidnow->error * pidnow->Ki;
-    if (pidnow->integral > pidnow->integral_limit) {
-        pidnow->integral = pidnow->integral_limit;
-    }
-    if (pidnow->integral < -pidnow->integral_limit) {
-        pidnow->integral = -pidnow->integral_limit;
-    }
-
-    pidnow->derivative = pidnow->Kd * (pidnow->error - pidnow->last_error);
-    pidnow->last_error = pidnow->error;
-
-    pidnow->pid_output = pidnow->error * pidnow->Kp + pidnow->integral + pidnow->derivative;
-
-    if (pidnow->pid_output > pidnow->output_limit) {
-        pidnow->pid_output = pidnow->output_limit;
-    }
-    if (pidnow->pid_output < -pidnow->output_limit) {
-        pidnow->pid_output = -pidnow->output_limit;
-    }
-    return pidnow->pid_output;
-}
 
 void loadEEpromSettings()
 {
@@ -832,98 +812,8 @@ void saveEEpromSettings()
     save_flash_nolib(eepromBuffer.buffer, sizeof(eepromBuffer.buffer), eeprom_address);
 }
 
-uint16_t getSmoothedCurrent()
-{
-    total = total - readings[readIndex];
-    readings[readIndex] = ADC_raw_current;
-    total = total + readings[readIndex];
-    readIndex = readIndex + 1;
-    if (readIndex >= numReadings) {
-        readIndex = 0;
-    }
-    smoothedcurrent = total / numReadings;
-    return smoothedcurrent;
-}
 
-RAM_FUNC void getBemfState()
-{
-    uint8_t current_state = 0;
-#if defined(MCU_F031) || defined(MCU_G031)
-    if (step == 1 || step == 4) {
-        current_state = PHASE_C_EXTI_PORT->IDR & PHASE_C_EXTI_PIN;
-    }
-    if (step == 2 || step == 5) { //        in phase two or 5 read from phase A Pf1
-        current_state = PHASE_A_EXTI_PORT->IDR & PHASE_A_EXTI_PIN;
-    }
-    if (step == 3 || step == 6) { // phase B pf0
-        current_state = PHASE_B_EXTI_PORT->IDR & PHASE_B_EXTI_PIN;
-    }
-#else
-    //Get current comparator output level
-    current_state = !getCompOutputLevel(); // polarity reversed
-#endif
-    if (rising) {
-        if (current_state) {
-            bemfcounter++;
-        } else {
-            bad_count++;
-            if (bad_count > bad_count_threshold) {
-                bemfcounter = 0;
-            }
-        }
-    } else {
-        if (!current_state) {
-            bemfcounter++;
-        } else {
-            bad_count++;
-            if (bad_count > bad_count_threshold) {
-                bemfcounter = 0;
-            }
-        }
-    }
-}
 
-RAM_FUNC void commutate()
-{
-    if (forward == 1) {
-        step++;
-        if (step > 6) {
-            step = 1;
-            desync_check = 1;
-        }
-        rising = step % 2;
-    } else {
-        step--;
-        if (step < 1) {
-            step = 6;
-            desync_check = 1;
-        }
-        rising = !(step % 2);
-    }
-#ifdef INVERTED_EXTI
-    rising = !rising;
-#endif
-    __disable_irq(); // don't let dshot interrupt
-    if (!prop_brake_active) {
-        comStep(step);
-    }
-    __enable_irq();
-    changeCompInput();
-#ifndef NO_POLLING_START
-	if (average_interval > polling_mode_changeover + 500) {
-      old_routine = 1;
-   }
-#endif
-    bemfcounter = 0;
-    zcfound = 0;
-    commutation_intervals[step - 1] = commutation_interval; // just used to calulate average
-    
-#ifdef USE_PULSE_OUT
-	if(step == 1 || step == 4  ){
-    WRITE_REG(RPM_PULSE_PORT->ODR, READ_REG(RPM_PULSE_PORT->ODR) ^ RPM_PULSE_PIN);
-	}
-#endif
-}
 
 /*
  * @brief 	Called by the COM_TIMER interrupt handler after the set wait time
@@ -931,858 +821,18 @@ RAM_FUNC void commutate()
  * 			This disables the COM_TIMER interrupt.
  * 			Then it enables the comparator to generate its interrupt.
  */
-RAM_FUNC void PeriodElapsedCallback()
-{
-    DISABLE_COM_TIMER_INT(); // disable interrupt
-    commutate();
-    commutation_interval = ((commutation_interval)+((lastzctime + thiszctime) >> 1))>>1;
-  	if (!eepromBuffer.auto_advance) {
-	  advance = (commutation_interval * temp_advance) >> 6; // 60 divde 64 0.9375 degree increments
-	} else {
-	  advance = (commutation_interval * auto_advance_level) >> 6; // 60 divde 64 0.9375 degree increments
-    }
-    waitTime = (commutation_interval >> 1) - advance;
-    if (!old_routine) {
-        enableCompInterrupts(); // enable comp interrupt
-    }
-    if (zero_crosses < 10000) {
-        zero_crosses++;
-    }
-    HWCI_PERF_ZC();
-}
 
 /*
  * @brief 	Called by the comparator interrupt handler.
  * 			Disables the comparator interrupt.
  * 			Enables the COM_TIMER and sets it to generate an interrupt after the wait time.
  */
-RAM_FUNC void interruptRoutine()
-{
-    HWCI_PERF_ZC_PHASE_CAPTURE(); // TIM1 phase at ISR entry, pre-confirm
-#ifdef MCU_F051
-    // PWM phase (TIM1 CNT) at the edge, sampled before the confirm loop's
-    // wall-clock window advances it: the turn-on-pileup compensation below
-    // keys off where the comparator edge actually appeared, not where it was
-    // finally accepted.
-    uint16_t zc_pwm_cnt = (uint16_t)TIM1->CNT;
-#endif
-//   if (average_interval > 125) {
-//        if ((INTERVAL_TIMER_COUNT < 125) && (duty_cycle < 600) && (zero_crosses < 500)) { // should be impossible, desync?exit anyway
-//           return;
-//        }
-//        stuckcounter++; // stuck at 100 interrupts before the main loop happens
-//                        // again.
-//        if (stuckcounter > 100) {
-//            maskPhaseInterrupts();
-//            zero_crosses = 0;
-//            return;
-//        }
-//    }
-        // Zero-cross confirm: reject unless the window's reads hold the
-        // post-crossing level. Loop speed sets the sampling window: inlining
-        // getCompOutputLevel removes the per-sample call overhead, and with the
-        // companion RAM-execution change the loop is faster still; filter_level
-        // is scaled (42/10/7 on F051) to keep the same wall-clock window as the
-        // stock ~56-cycle sampling cadence.
-#ifdef MCU_F051
-        // Glitch-tolerant variant: the ~16-cycle cadence lands on a brief
-        // comparator glitch ~3x more often than stock sampling, and a strict
-        // all-samples-must-agree confirm defers detection to the NEXT
-        // comparator edge - up to a PWM period late. Bench-measured on the
-        // ARK 4IN1: 2-3x higher commutation jitter at 15-20 kHz commutation
-        // rates vs stock cadence (upstream 2.1% vs 5.4% at full throttle).
-        // Tolerating up to filter_level/4 bad samples per window accepts
-        // through isolated glitches while a genuinely un-crossed level still
-        // rejects via the early-out; the full window length (and so the
-        // sustained-noise immunity of the filter_level retune) is unchanged.
-        {
-            int bad = 0;
-            const int tolerance = filter_level >> 2;
-            for (int i = 0; i < filter_level; i++) {
-                if (getCompOutputLevel() == rising) {
-                    if (++bad > tolerance) {
-                        HWCI_PERF_CONFIRM_REJECT();
-                        return;
-                    }
-                }
-            }
-        }
-#else
-        for (int i = 0; i < filter_level; i++) {
-#if defined(MCU_F031) || defined(MCU_G031)
-            if (((current_GPIO_PORT->IDR & current_GPIO_PIN) == !(rising))) {
-#else
-            if (getCompOutputLevel() == rising) {
-#endif
-                HWCI_PERF_CONFIRM_REJECT();
-                return;
-            }
-        }
-#endif
-#ifdef MCU_F051
-    // Turn-on-pileup timestamp compensation. A zero-cross that physically
-    // occurs during the PWM off-window (freewheel) is invisible to the
-    // comparator and is only registered at the next turn-on, quantizing the
-    // timestamp to the PWM grid; those corrupted intervals feed waitTime and
-    // the commutation loop hunts, amplifying the quantization into the
-    // measured jitter hump (bench: ~23% at t60/t70 where the commutation rate
-    // beats against the 24 kHz PWM). Detections of such crossings land 1-3
-    // phase bins after turn-on (bin 1 peak = comparator+ISR latency of
-    // ~1.3-2.6 us; bin 0 stays at the uniform rate) and the true crossing lay
-    // on average half the off-window earlier - back-date thiszctime by that
-    // half-window so the loop tracks the estimated true crossing instead of
-    // the grid position.
-    //
-    // Guards, each tied to a bench-observed failure of the unguarded version
-    // (which cut hump jitter 23.4->17.9% but desynced on a throttle step and
-    // doubled low-RPM jitter):
-    //  - correct thiszctime ONLY; the interval timer still resets to zero at
-    //    the detection instant, so one bad estimate perturbs one interval and
-    //    cannot cascade into the next (the cascading variant lost sync at 70A)
-    //  - duty-slew guard: skip while adjusted_duty_cycle is moving, i.e.
-    //    during throttle transients, where the loop must track real
-    //    acceleration rather than have timestamps rewritten under it
-    //  - hump band only (commutation interval < 2 PWM periods): at low RPM the
-    //    interval dwarfs the PWM period, there is no hunting to break, and the
-    //    correction only adds variance
-    //  - clamp to interval/8 as a backstop against any degenerate state
-    // Units: INTERVAL_TIMER (TIM2) is 2 MHz and TIM1 is 48 MHz, so one INTERVAL
-    // tick is 24 TIM1 ticks; half an off-window of N TIM1 ticks is N/48 INTERVAL
-    // ticks, computed as (N * 1365) >> 16 to keep a soft divide out of the ISR.
-    uint16_t zc_grid_comp = 0;
-    {
-        static uint16_t zc_prev_duty;
-        const uint16_t zc_arr = tim1_arr;
-        const uint16_t zc_duty = adjusted_duty_cycle;
-        const uint16_t zc_slew = (zc_duty >= zc_prev_duty)
-            ? (uint16_t)(zc_duty - zc_prev_duty)
-            : (uint16_t)(zc_prev_duty - zc_duty);
-        zc_prev_duty = zc_duty;
-        const uint32_t zc_ci = commutation_interval;
-        if (zero_crosses >= 100
-                && zc_duty < zc_arr                      /* off-window exists */
-                && zc_pwm_cnt >= (uint16_t)(zc_arr >> 5) /* pile-up bins 1-3 */
-                && zc_pwm_cnt < (uint16_t)(zc_arr >> 3)
-                && zc_slew <= (uint16_t)(zc_arr >> 8)    /* duty steady */
-                && (zc_ci * 12u) < (uint32_t)zc_arr + 1u /* hump band only */
-        ) {
-            uint32_t comp = ((uint32_t)(zc_arr - zc_duty) * 1365u) >> 16;
-            const uint32_t cap = zc_ci >> 3;
-            if (comp > cap) {
-                comp = cap;
-            }
-            zc_grid_comp = (uint16_t)comp;
-        }
-    }
-#endif
-    __disable_irq();
-    maskPhaseInterrupts();
-    lastzctime = thiszctime;
-#ifdef MCU_F051
-    thiszctime = (uint16_t)(INTERVAL_TIMER_COUNT - zc_grid_comp);
-#else
-    thiszctime = INTERVAL_TIMER_COUNT;
-#endif
-    SET_INTERVAL_TIMER_COUNT(0);
-    SET_AND_ENABLE_COM_INT(waitTime+1); // enable COM_TIMER interrupt
-    __enable_irq();
-    HWCI_PERF_ZC_PHASE_COMMIT(); // accepted edge: bin its PWM phase
-}
 
-void startMotor()
-{
-    if (running == 0) {
-        commutate();
-        commutation_interval = 10000;
-        SET_INTERVAL_TIMER_COUNT(5000);
-        running = 1;
-    }
-    enableCompInterrupts();
-}
 
-void setInput()
-{
-    if (eepromBuffer.bi_direction) {
-        if (dshot == 0) {
-            if (eepromBuffer.rc_car_reverse) {
-                if (newinput > (1000 + (servo_dead_band << 1))) {
-                    if (forward == eepromBuffer.dir_reversed) {
-                        adjusted_input = 0;
-                        //               if (running) {
-                        prop_brake_active = 1;
-                        if (return_to_center) {
-                            forward = 1 - eepromBuffer.dir_reversed;
-                            prop_brake_active = 0;
-                            return_to_center = 0;
-                        }
-                    }
-                    if (prop_brake_active == 0) {
-                        return_to_center = 0;
-                        adjusted_input = map(newinput, 1000 + (servo_dead_band << 1), 2000, 47, 2047);
-                    }
-                }
-                if (newinput < (1000 - (servo_dead_band << 1))) {
-                    if (forward == (1 - eepromBuffer.dir_reversed)) {
-                        adjusted_input = 0;
-                        prop_brake_active = 1;
-                        if (return_to_center) {
-                            forward = eepromBuffer.dir_reversed;
-                            prop_brake_active = 0;
-                            return_to_center = 0;
-                        }
-                    }
-                    if (prop_brake_active == 0) {
-                        return_to_center = 0;
-                        adjusted_input = map(newinput, 0, 1000 - (servo_dead_band << 1), 2047, 47);
-                    }
-                }
-                if (newinput >= (1000 - (servo_dead_band << 1)) && newinput <= (1000 + (servo_dead_band << 1))) {
-                    adjusted_input = 0;
-                    if (prop_brake_active) {
-                        prop_brake_active = 0;
-                        return_to_center = 1;
-                    }
-                }
-            } else {
-                if (newinput > (1000 + (servo_dead_band << 1))) {
-                    if (forward == eepromBuffer.dir_reversed) {
-                        if (((commutation_interval > reverse_speed_threshold) && (duty_cycle < 200)) || stepper_sine) {
-                            forward = 1 - eepromBuffer.dir_reversed;
-                            zero_crosses = 0;
-                            old_routine = 1;
-                            maskPhaseInterrupts();
-                            brushed_direction_set = 0;
-                        } else {
-                            newinput = 1000;
-                        }
-                    }
-                    adjusted_input = map(newinput, 1000 + (servo_dead_band << 1), 2000, 47, 2047);
-                }
-                if (newinput < (1000 - (servo_dead_band << 1))) {
-                    if (forward == (1 - eepromBuffer.dir_reversed)) {
-                        if (((commutation_interval > reverse_speed_threshold) && (duty_cycle < 200)) || stepper_sine) {
-                            zero_crosses = 0;
-                            old_routine = 1;
-                            forward = eepromBuffer.dir_reversed;
-                            maskPhaseInterrupts();
-                            brushed_direction_set = 0;
-                        } else {
-                            newinput = 1000;
-                        }
-                    }
-                    adjusted_input = map(newinput, 0, 1000 - (servo_dead_band << 1), 2047, 47);
-                }
 
-                if (newinput >= (1000 - (servo_dead_band << 1)) && newinput <= (1000 + (servo_dead_band << 1))) {
-                    adjusted_input = 0;
-                    brushed_direction_set = 0;
-                }
-            }
-        }
-        if (dshot) {
-                     if (eepromBuffer.rc_car_reverse) {
-                         if (newinput > 1047) {
-                         if (forward == eepromBuffer.dir_reversed) {
-                         adjusted_input = 0;
-                         prop_brake_active = 1;
-                         if (return_to_center) {
-                             forward = 1 - eepromBuffer.dir_reversed;
-                             prop_brake_active = 0;
-                             return_to_center = 0;
-                         }
-                     }
-                     if (prop_brake_active == 0) {
-                         return_to_center = 0;
-                         adjusted_input = ((newinput - 1048) * 2 + 47) - reversing_dead_band;
-                     }
-                     }
-                     if (newinput <= 1047 && newinput > 47) {
-                     if (forward == (1 - eepromBuffer.dir_reversed)) {
-                         adjusted_input = 0;
-                         prop_brake_active = 1;
-                         if (return_to_center) {
-                             forward = eepromBuffer.dir_reversed;
-                             prop_brake_active = 0;
-                             return_to_center = 0;
-                         }
-                     }
-                     if (prop_brake_active == 0) {
-                         return_to_center = 0;
-                         adjusted_input = ((newinput - 48) * 2 + 47) - reversing_dead_band;
-                     }
-                     }
-                     if (newinput < 48) {
-                     adjusted_input = 0;
-                     if (prop_brake_active) {
-                         prop_brake_active = 0;
-                         return_to_center = 1;
-                     }
-                 }
-                         } else {
-            if (newinput > 1047) {
 
-                if (forward == eepromBuffer.dir_reversed) {
-                    if (((commutation_interval > reverse_speed_threshold) && (duty_cycle < 200)) || stepper_sine) {
-                        forward = 1 - eepromBuffer.dir_reversed;
-                        zero_crosses = 0;
-                        old_routine = 1;
-                        maskPhaseInterrupts();
-                        brushed_direction_set = 0;
-                     } else {
-                        newinput = 0;
-                    }
-                }
-                adjusted_input = ((newinput - 1048) * 2 + 47) - reversing_dead_band;
-            }
-            if (newinput <= 1047 && newinput > 47) {
-                if (forward == (1 - eepromBuffer.dir_reversed)) {
-                    if (((commutation_interval > reverse_speed_threshold) && (duty_cycle < 200)) || stepper_sine) {
-                        zero_crosses = 0;
-                        old_routine = 1;
-                        forward = eepromBuffer.dir_reversed;
-                        maskPhaseInterrupts();
-                        brushed_direction_set = 0;
-                     } else {
-                        newinput = 0;
-                    }
-                }
-                adjusted_input = ((newinput - 48) * 2 + 47) - reversing_dead_band;
-            }
-            if (newinput < 48) {
-                adjusted_input = 0;
-                brushed_direction_set = 0;
-                }
-            }
-        }
-    } else {
-        adjusted_input = newinput;
-    }
-#ifndef BRUSHED_MODE
-    if ((bemf_timeout_happened > bemf_timeout) && eepromBuffer.stuck_rotor_protection) {
-        allOff();
-        maskPhaseInterrupts();
-        input = 0;
-        bemf_timeout_happened = 102;
-#ifdef USE_RGB_LED
-        setIndividualRGBLed(1, 0, 0);
-#endif
-    } else {
-#ifdef FIXED_DUTY_MODE
-        input = FIXED_DUTY_MODE_POWER * 20 + 47;
-#else
-        if (eepromBuffer.use_sine_start) {
-            if (adjusted_input < 30) { // dead band ?
-                input = 0;
-            }
-            if (adjusted_input > 30 && adjusted_input < (eepromBuffer.sine_mode_changeover_thottle_level * 20)) {
-                input = map(adjusted_input, 30,
-                    (eepromBuffer.sine_mode_changeover_thottle_level * 20), 47, 160);
-            }
-            if (adjusted_input >= (eepromBuffer.sine_mode_changeover_thottle_level * 20)) {
-                input = map(adjusted_input, (eepromBuffer.sine_mode_changeover_thottle_level * 20),
-                    2047, 160, 2047);
-            }
-        } else {
-            if (use_speed_control_loop) {
-                if (drive_by_rpm) {
-                    target_e_com_time = 60000000 / map(adjusted_input, 47, 2047, MINIMUM_RPM_SPEED_CONTROL, MAXIMUM_RPM_SPEED_CONTROL) / (eepromBuffer.motor_poles / 2);
-                    if (adjusted_input < 47) { // dead band ?
-                        input = 0;
-                        speedPid.error = 0;
-                        input_override = 0;
-                    } else {
-                        input = (uint16_t)(input_override / 10000); // speed control pid override
-                        if (input > 2047) {
-                            input = 2047;
-                        }
-                        if (input < 48) {
-                            input = 48;
-                        }
-                    }
-                } else {
 
-                    input = (uint16_t)(input_override / 10000); // speed control pid override
-                    if (input > 2047) {
-                        input = 2047;
-                    }
-                    if (input < 48) {
-                        input = 48;
-                    }
-                }
-            } else {
 
-                input = adjusted_input;
-            }
-        }
-#endif
-    }
-#endif
-#ifndef BRUSHED_MODE
-if (!stepper_sine && armed) {
-        if (input >= 47 + (80 * eepromBuffer.use_sine_start)) {
-            if (running == 0) {
-                allOff();
-                if (!old_routine) {
-                    startMotor();
-                }
-                running = 1;
-                last_duty_cycle = min_startup_duty;
-            }
-
-            // straight line from (in_min, out_min) to (2047, 2000) using a
-            // startup computed Q16 slope, avoids calling map() at input rate
-            if (eepromBuffer.use_sine_start) {
-                duty_cycle_setpoint = input >= 2047 ? 2000
-                    : input <= 137 ? minimum_duty_cycle + 40
-                    : minimum_duty_cycle + 40 + (uint16_t)(((uint32_t)(input - 137) * sine_throttle_duty_slope_q16) >> 16);
-            } else {
-                duty_cycle_setpoint = input >= 2047 ? 2000
-                    : input <= 47 ? minimum_duty_cycle
-                    : minimum_duty_cycle + (uint16_t)(((uint32_t)(input - 47) * throttle_duty_slope_q16) >> 16);
-            }
-
-            if (!eepromBuffer.rc_car_reverse) {
-                prop_brake_active = 0;
-            }
-        }
-
-        if (input < 47 + (80 * eepromBuffer.use_sine_start)) {
-            if (play_tone_flag != 0) {
-                switch (play_tone_flag) {
-									
-                case 1:
-                    playDefaultTone();
-                    break;
-                case 2:
-                    playChangedTone();
-                    break;
-                case 3:
-                    playBeaconTune3();
-                    break;
-                case 4:
-                    playInputTune2();
-                    break;
-                case 5:
-                    playDefaultTone();
-                    break;
-                }
-                play_tone_flag = 0;
-            }
-
-            if (!eepromBuffer.comp_pwm) {
-                duty_cycle_setpoint = 0;
-                if (!running) {
-                    old_routine = 1;
-                    zero_crosses = 0;
-                    if (eepromBuffer.brake_on_stop) {
-                        fullBrake();
-                    } else {
-                        if (!prop_brake_active) {
-                            allOff();
-                        }
-                    }
-                }
-                if (eepromBuffer.rc_car_reverse && prop_brake_active) {
-#ifndef PWM_ENABLE_BRIDGE
-
-                  if (dshot == 0) prop_brake_duty_cycle = (getAbsDif(1000, newinput) + 1000);
-                    if (dshot)  {
-                        if (newinput <= 1047 && newinput > 47) prop_brake_duty_cycle = ((newinput - 48) * 2 + 47) - reversing_dead_band;
-                        if (newinput > 1047) prop_brake_duty_cycle = ((newinput - 1048) * 2 + 47) - reversing_dead_band;
-                    }
-                    if (prop_brake_duty_cycle >= (1999)) {
-
-                        fullBrake();
-                    } else {
-                        proportionalBrake();
-                    }
-#endif
-                }
-            } else {
-                if (!running) {
-
-                    old_routine = 1;
-                    zero_crosses = 0;
-                    bad_count = 0;
-                    if (eepromBuffer.brake_on_stop > 0) {
-                        if (!eepromBuffer.use_sine_start) {
-#ifndef PWM_ENABLE_BRIDGE
-                          if(eepromBuffer.brake_on_stop == 1){
-                             prop_brake_duty_cycle =  eepromBuffer.drag_brake_strength * 200;
-                              if (prop_brake_duty_cycle >= (1999)) {
-                                fullBrake();
-                              } else {
-                                proportionalBrake();
-                                prop_brake_active = 1;
-                              }
-                           }
-#else
-                            // todo add proportional braking for pwm/enable style bridge.
-#endif
-                        }
-                    } else {
-                        allOff();
-                    }
-                    duty_cycle_setpoint = 0;
-                }
-
-                phase_A_position = ((step - 1) * 60) + enter_sine_angle;
-                if (phase_A_position > 359) {
-                    phase_A_position -= 360;
-                }
-                phase_B_position = phase_A_position + 119;
-                if (phase_B_position > 359) {
-                    phase_B_position -= 360;
-                }
-                phase_C_position = phase_A_position + 239;
-                if (phase_C_position > 359) {
-                    phase_C_position -= 360;
-                }
-
-                if (eepromBuffer.use_sine_start == 1) {
-                    stepper_sine = 1;
-                }
-                duty_cycle_setpoint = 0;
-            }
-        }
-        if (!prop_brake_active) {
-            if (input >= 47 && (zero_crosses < (uint32_t)(30 >> eepromBuffer.stall_protection))) {
-                if (duty_cycle_setpoint < min_startup_duty) {
-                    duty_cycle_setpoint = min_startup_duty;
-                }
-                if (duty_cycle_setpoint > startup_max_duty_cycle) {
-                    duty_cycle_setpoint = startup_max_duty_cycle;
-                }
-            }
-
-            if (duty_cycle_setpoint > duty_cycle_maximum) {
-                duty_cycle_setpoint = duty_cycle_maximum;
-            }
-            if (use_current_limit) {
-                if (duty_cycle_setpoint > use_current_limit_adjust) {
-                    duty_cycle_setpoint = use_current_limit_adjust;
-                }
-            }
-
-            if (stall_protection_adjust > 0 && input > 47) {
-
-                duty_cycle_setpoint = duty_cycle_setpoint + (uint16_t)(stall_protection_adjust/10000);
-            }
-        }
-    }
-#endif
-}
-
-RAM_FUNC void tenKhzRoutine()
-{ // 20khz as of 2.00 to be renamed
-    HWCI_PERF_CTRL_ENTER();
-    duty_cycle = duty_cycle_setpoint;
-    tenkhzcounter++;
-    ledcounter++;
-    ramp_count++;
-    one_khz_loop_counter++;
-    if (!armed) {
-        if (cell_count == 0) {
-            if (inputSet) {
-                if (adjusted_input == 0) {
-                    armed_timeout_count++;
-                    if (armed_timeout_count > LOOP_FREQUENCY_HZ) { // one second
-                        if (zero_input_count > 30) {
-                            armed = 1;
-#ifdef USE_LED_STRIP
-                            //	send_LED_RGB(0,0,0);
-                            delayMicros(1000);
-                            send_LED_RGB(0, 255, 0);
-#endif
-#ifdef USE_RGB_LED
-                            setIndividualRGBLed(0,1,0);
-#endif
-                            if ((cell_count == 0) && eepromBuffer.low_voltage_cut_off == 1) {
-                                cell_count = battery_voltage / 370;
-                                for (int i = 0; i < cell_count; i++) {
-                                    playInputTune();
-                                    delayMillis(100);
-                                    RELOAD_WATCHDOG_COUNTER();
-                                }
-                            } else {
-#ifdef MCU_AT415
-															play_tone_flag = 4;
-#else
-															playInputTune();
-#endif
-                            }
-                            if (!servoPwm && !dshot) {
-                                eepromBuffer.rc_car_reverse = 0;
-                            }
-                        } else {
-                            inputSet = 0;
-                            armed_timeout_count = 0;
-                        }
-                    }
-                } else {
-                    armed_timeout_count = 0;
-                }
-            }
-        }
-    }
-
-    if (eepromBuffer.telemetry_on_interval) {
-        telem_ms_count++;
-        if (telem_ms_count > ((telemetry_interval_ms - 1 + eepromBuffer.telemetry_on_interval) * 20)) {
-            // telemetry_on_interval = 1 is a boolean, but it can also be 2 or more to indicate an identifier
-            // by making the interval just slightly different with an unique identifier, we can guarantee that many ESCs can communicate on just one signal
-            // there will be some collisions but not as many as if two ESCs always tried to talk at once.
-            send_telemetry = 1;
-            telem_ms_count = 0;
-        }
-    }
-
-#ifndef BRUSHED_MODE
-
-    if (!stepper_sine) {
-#ifndef CUSTOM_RAMP
-        if (old_routine && running) {
-	//				send_LED_RGB(255, 0, 0);
-            maskPhaseInterrupts();
-            getBemfState();
-            if (!zcfound) {
-                if (rising) {
-                    if (bemfcounter > min_bemf_counts_up) {
-                        zcfound = 1;
-                        zcfoundroutine();
-                    }
-                } else {
-                    if (bemfcounter > min_bemf_counts_down) {
-                        zcfound = 1;
-                        zcfoundroutine();
-                    }
-                }
-            }
-        }
-#endif
-        if (one_khz_loop_counter > PID_LOOP_DIVIDER) { // 1khz PID loop
-            PROCESS_ADC_FLAG = 1; // set flag to do new adc read at lower priority
-            one_khz_loop_counter = 0;
-            if (use_current_limit && running) {
-                use_current_limit_adjust -= (int16_t)(doPidCalculations(&currentPid, actual_current,
-                                                          eepromBuffer.limits.current * 2 * 100)
-                    / 10000);
-                if (use_current_limit_adjust < minimum_duty_cycle) {
-                    use_current_limit_adjust = minimum_duty_cycle;
-                }
-                if (use_current_limit_adjust > 2000) {
-                    use_current_limit_adjust = 2000;
-                }
-            }
-            if (eepromBuffer.stall_protection && running) { // this boosts throttle as the rpm gets lower, for crawlers
-                                               // and rc cars only, do not use for multirotors.
-                stall_protection_adjust += (doPidCalculations(&stallPid, commutation_interval,
-                                               stall_protect_target_interval));
-                if (stall_protection_adjust > 150 * 10000) {
-                    stall_protection_adjust = 150 * 10000;
-                }
-                if (stall_protection_adjust <= 0) {
-                    stall_protection_adjust = 0;
-                }
-            }
-            if (use_speed_control_loop && running) {
-                input_override += doPidCalculations(&speedPid, e_com_time, target_e_com_time);
-                if (input_override > 2047 * 10000) {
-                    input_override = 2047 * 10000;
-                }
-                if (input_override < 0) {
-                    input_override = 0;
-                }
-                if (zero_crosses < 100) {
-                    speedPid.integral = 0;
-                }
-            }
-        }
-        if (ramp_count > ramp_divider) {
-          ramp_count = 0;
-#ifdef VOLTAGE_BASED_RAMP
-            uint16_t voltage_based_max_change = map(battery_voltage, 800, 2200, 10, 1);
-            if (average_interval > 200) {
-                max_duty_cycle_change = voltage_based_max_change;
-            } else {
-                max_duty_cycle_change = voltage_based_max_change * 3;
-            }
-#else
-            if (zero_crosses < 150 || last_duty_cycle < 150) {   
-                max_duty_cycle_change = max_ramp_startup;
-            } else {
-                if (average_interval > 500) {
-                    max_duty_cycle_change = max_ramp_low_rpm;
-                } else {
-                    max_duty_cycle_change = max_ramp_high_rpm;
-                }
-            }
-          
-#endif
-#ifdef CUSTOM_RAMP
-   //         max_duty_cycle_change = eepromBuffer[30];
-#endif
-            if ((duty_cycle - last_duty_cycle) > max_duty_cycle_change) {
-                duty_cycle = last_duty_cycle + max_duty_cycle_change;
-
-            }
-            if ((last_duty_cycle - duty_cycle) > max_duty_cycle_change) {
-                duty_cycle = last_duty_cycle - max_duty_cycle_change;
-            }
-            }else{
-             duty_cycle = last_duty_cycle;
-            }
-
-        if ((armed && running) && input > 47) {
-            if (eepromBuffer.variable_pwm) {
-            }
-            adjusted_duty_cycle = (((uint32_t)duty_cycle * pwm_to_arr_scale_q16) >> 16) + 1;
-
-        } else {
-
-            if (prop_brake_active) {
-              adjusted_duty_cycle =  tim1_arr - (((uint32_t)prop_brake_duty_cycle * pwm_to_arr_scale_q16) >> 16);
-            } else {
-              if((eepromBuffer.brake_on_stop == 2) && armed){  // require arming for active brake
-                comStep(2);
-                adjusted_duty_cycle = DEAD_TIME + (((uint32_t)eepromBuffer.active_brake_power * pwm_to_arr_scale_q16) >> 16)* 10;
-            }else{
-                adjusted_duty_cycle = (((uint32_t)duty_cycle * pwm_to_arr_scale_q16) >> 16);
-            }
-            }
-        }
-        last_duty_cycle = duty_cycle;
-        SET_AUTO_RELOAD_PWM(tim1_arr);
-        SET_DUTY_CYCLE_ALL(adjusted_duty_cycle);
-    }
-#endif // ndef brushed_mode
-#if defined(FIXED_DUTY_MODE) || defined(FIXED_SPEED_MODE)
-    if (getInputPinState()) {
-        signaltimeout++;
-        if (signaltimeout > LOOP_FREQUENCY_HZ) {
-            NVIC_SystemReset();
-        }
-    } else {
-        signaltimeout = 0;
-    }
-#else
-    signaltimeout++;
-
-#endif
-    HWCI_PERF_CTRL_EXIT();
-}
-
-void processDshot()
-{
-    if (compute_dshot_flag == 1) {
-        computeDshotDMA();
-        compute_dshot_flag = 0;
-    }
-    if (compute_dshot_flag == 2) {
-      if(e_com_time > 65535){    // beyond dshot range
-        make_dshot_package(65535);
-      }else{
-        make_dshot_package(e_com_time);
-      }
-        compute_dshot_flag = 0;
-        return;
-    }
-    setInput();
-}
-
-void advanceincrement()
-{
-    if (!forward) {
-        phase_A_position++;
-        if (phase_A_position > 359) {
-            phase_A_position = 0;
-        }
-        phase_B_position++;
-        if (phase_B_position > 359) {
-            phase_B_position = 0;
-        }
-        phase_C_position++;
-        if (phase_C_position > 359) {
-            phase_C_position = 0;
-        }
-    } else {
-        phase_A_position--;
-        if (phase_A_position < 0) {
-            phase_A_position = 359;
-        }
-        phase_B_position--;
-        if (phase_B_position < 0) {
-            phase_B_position = 359;
-        }
-        phase_C_position--;
-        if (phase_C_position < 0) {
-            phase_C_position = 359;
-        }
-    }
-#ifdef GIMBAL_MODE
-    setPWMCompare1(((2 * pwmSin[phase_A_position]) + gate_drive_offset) * TIMER1_MAX_ARR / 2000);
-    setPWMCompare2(((2 * pwmSin[phase_B_position]) + gate_drive_offset) * TIMER1_MAX_ARR / 2000);
-    setPWMCompare3(((2 * pwmSin[phase_C_position]) + gate_drive_offset) * TIMER1_MAX_ARR / 2000);
-#else
-    setPWMCompare1(
-        (((2 * pwmSin[phase_A_position] / SINE_DIVIDER) + gate_drive_offset) * TIMER1_MAX_ARR / 2000) * eepromBuffer.sine_mode_power / 10);
-    setPWMCompare2(
-        (((2 * pwmSin[phase_B_position] / SINE_DIVIDER) + gate_drive_offset) * TIMER1_MAX_ARR / 2000) * eepromBuffer.sine_mode_power / 10);
-    setPWMCompare3(
-        (((2 * pwmSin[phase_C_position] / SINE_DIVIDER) + gate_drive_offset) * TIMER1_MAX_ARR / 2000) * eepromBuffer.sine_mode_power / 10);
-#endif
-}
-
-void zcfoundroutine()
-{ // only used in polling mode, blocking routine.
-    thiszctime = INTERVAL_TIMER_COUNT;
-    SET_INTERVAL_TIMER_COUNT(0);
-    commutation_interval = (thiszctime + (3 * commutation_interval)) / 4;
-    advance = (temp_advance * commutation_interval) >> 6; //   7.5 degree increments
-    waitTime = commutation_interval / 2 - advance;
-    while ((INTERVAL_TIMER_COUNT) < (waitTime)) {
-        if (zero_crosses < 5) {
-            break;
-        }
-    }
-#ifdef MCU_GDE23
-    TIMER_CAR(COM_TIMER) = waitTime;
-#endif
-#ifdef STMICRO
-    COM_TIMER->ARR = waitTime;
-#endif
-#ifdef MCU_AT32
-		COM_TIMER->pr = waitTime;
-#endif
-#ifdef NXP
-//	COM_TIMER->MSR[0] = waitTime;
-	COM_TIMER->MR[0] = waitTime;
-#endif
-
-    commutate();
-    bemfcounter = 0;
-    bad_count = 0;
-
-    zero_crosses++;
-#ifdef NO_POLLING_START     // changes to interrupt mode after 2 zero crosses, does not re-enter
-       if (zero_crosses > 2) {
-            old_routine = 0;
-            enableCompInterrupts(); // enable interrupt
-        }
-#else
-    if (eepromBuffer.stall_protection || eepromBuffer.rc_car_reverse) {
-        if (zero_crosses >= 20 && commutation_interval <= 2000) {
-            old_routine = 0;
-            enableCompInterrupts(); // enable interrupt
-        }
-    } else {
-       if (commutation_interval < polling_mode_changeover) {
-            old_routine = 0;
-            enableCompInterrupts(); // enable interrupt
-        }
-    }
- #endif
-}
 #ifdef BRUSHED_MODE
 void runBrushedLoop()
 {
@@ -1846,7 +896,7 @@ void runBrushedLoop()
 /*
   check device info from the bootloader, confirming pin code and eeprom location
  */
-static void checkDeviceInfo(void)
+static void __attribute__((noinline)) checkDeviceInfo(void)
 {
 #ifdef MCU_SITL
     // no bootloader device info page in SITL
@@ -1865,18 +915,31 @@ static void checkDeviceInfo(void)
 #define DEVINFO_MAGIC1 0x5925e3da
 #define DEVINFO_MAGIC2 0x4eb863d9
 
-    const struct devinfo {
+    // Fixed bootloader address; GCC 12+ -Warray-bounds needs care.
+    struct devinfo {
         uint32_t magic1;
         uint32_t magic2;
-        const uint8_t deviceInfo[9];
-    } *devinfo = (struct devinfo *)(0x1000 - 32);
-    if (devinfo->magic1 != DEVINFO_MAGIC1 ||
-        devinfo->magic2 != DEVINFO_MAGIC2) {
+        uint8_t deviceInfo[9];
+    };
+    volatile const struct devinfo *devinfo =
+        (volatile const struct devinfo *)(uintptr_t)(0x1000u - 32u);
+#if defined(__GNUC__) && (__GNUC__ >= 12)
+#pragma GCC diagnostic push
+#pragma GCC diagnostic ignored "-Warray-bounds"
+#pragma GCC diagnostic ignored "-Wstringop-overread"
+#endif
+    const uint32_t magic1 = devinfo->magic1;
+    const uint32_t magic2 = devinfo->magic2;
+    const uint8_t eeprom_code = devinfo->deviceInfo[4];
+#if defined(__GNUC__) && (__GNUC__ >= 12)
+#pragma GCC diagnostic pop
+#endif
+    if (magic1 != DEVINFO_MAGIC1 || magic2 != DEVINFO_MAGIC2) {
         // bootloader does not support this feature, nothing to do
         return;
     }
     // change eeprom_address based on the code in the bootloaders device info
-    switch (devinfo->deviceInfo[4]) {
+    switch (eeprom_code) {
         case 0x1f:
             eeprom_address = 0x08007c00;
             break;
